Flesh out the design of server-client folder and message synchronization, in consideration of the IMAP protocol (RFC3501)
The synchronization is moving towards a simplified, 1-way server->client synchronization, which creates
local EmailInfo objects as a mirror of what is on the server. The intent is that the EmailInfo objects can
be deleted at will, and they do not store any state information which is not otherwise on the server.