Work-in-progress implementation of Email account settings
This initial implementation includes the data model definition and UI support to edit the information. More work is
needed to store the user name and password into the keychain, set a default port number when the "use SSL?"
setting is toggled, and  to keep rules and email information separated by account.