Refactor the PopupButtonListView class to be derived from a more generic ButtonListView
The ButtonListView can be a child of any view and doesn't by default include a cancel button to remove the view from its parent. Also removed
the SingleButtonTableFooterClass; it is no longer needed since a table footer with a single button can be represented by a
ButtonListView with a single button.