Move locally cached email information to app data database, reducing duplication and complexity in the code.
The "AppData" data model already supported cached email addresses. To support rules which match on the email
recipient, the EmailInfo object needs to have a field which points to the potentially multiple email addresses.
This was already supported in the AppData data model, but not in the EmailInfo data model. So the choice
was either to merge them and reduce the duplication and complexity, or keep them separate and add
even more duplication.