Restructure the data model to segregate domain, folder, address, and email information by account.
This includes creating references from the domain, folder, address and email objects back to the
parent account. This also includes changes to core data fetch parameters to use a predicate
which matches the current account. Finally, the message synchronization code has been
updated to set the account references when new objects are created.