Throw an exception if the IMAP server connection fails
With the integration of the updated MailCore 3rd party library, the connect method on the email account
returns a boolean of FALSE/NO if the connection fails rather than throwing an exception. The app's code
still expects an exception to be thrown. Without this change, the app assumed the connection took place
and continued with the synchronization, resulting in a crash.