If the user exits the message detail view before the message body is loaded, cancel the operation to load the message body.
The app was previously crashing if the message detail view was closed before the body was loaded.
The NSOperationQueue (thread queue) for loading message bodies has been moved to the AppDelegate,
and the NSOperation to retrieve the message body now checks the "isCancelled" property upon return
from any blocking IMAP server method.