Make sure the refresh button for IMAP account synchronizations updates even if the view is changed
Previously, if the user switched to another view while account synchronization was taking place, the
refresh button in the message view was not updating. By moving the registration of footer view as delegate to
receive notification of synchronization events to the viewDidLoad & viewDidUnload methods,
the refresh button will have the correct state even if the view is switched.