WIP implementation of refined design for saved message filters
This check-in includes eliminating the tabbed navigational controller for displaying the different message lists.
Instead, a single message list view controller is shown. At the top of the message list controller a settings button
is shown, where the user can manually edit saved rules/filters (previously in the tab controller).