Address a conflict with filters being reset in the flip view for message filter editing and the thread for counting the # of messages matching each thread.
The flip view for editing message filters has it's own NSManagedObjectContext (called 'editFilter' in EmailInfoTableViewController.m.
When changes are saved to this MOC, the changes are merged back to the main MOC, and the filter name is cleared if the current
filter was loaded from a saved one. The method responsible for merging the changes was previously launching a thread
in the background to also update the message counts matching each filter, so at that point there were 3 MOC's being edited,
including the main, one for the current filter, and the MOC for counting matching messages. The solution was to not launch
the counting operation when changes to the current filter were saved, thus keeping merges simple while the current filter
is being edited. These changes seem to have addressed the crash which was intermittently occurring when the
filter was reset in the flip view (but not the popup menu from the main view).