Handle the message deletion case where a message is moved to the delete folder while others are being deleted.
There's numerous comments in MailDeleteOperation explaining the implementation in detail. One
other improvement over the old code was to retrieve the next UID in the delete folder before
the deletions take place. After the delete, the list of messages after the next UID can be retrieved,
and these represent the list of new messages. Previously, an entire list of UIDs was being retrieved
before and after the messages were moved to the delete folder.