Incrementally improve the table views for selecting filters' domains and addresses.
The table view now has an index, the items are broken down into sections alphabetically,
and custom configuration of the UITableViewCell is used to ensure the information
is displayed more appropriately.

The domain list as is fairly usable, but the address list still has too may items to
select from. A custom table view for selecting addresses is needed.