Use a dedicated NSManagedObjectContext to create new email accounts.
Previously, EmailAccount objects were being created inside the main NSManagedObjectContext (wrapped by a DataModelController).
This caused conflicts if a partially-populated EmailInfo object was being created, then one of the other threads saved the
main object context. The solution was to use a dedicated NSManagedObjectContext, which ensures the creation of the account
can be canceled, will not be saved until it is fully populated, and will not conflict with the main NSManagedObjectContext.