Refine the synchronization algorithm to only download headers up to the maximum number of messages synchronized.
Previously, the algorithm to download message headers was download *all* the headers each time a message
synchronization took place. So, for example, if a user had 20K messages, the algorithm would download
all 20K headers each time, or however many hadn't been deleted yet. The number downloaded could easily
exceed the user's setting for maximum number of messages to sync at one time (which defaults to 10K)
The old algorithm was also based on message UID.

The new algorithm gets a total of messages across all synchronized mailboxes, then only downloads
a number of headers which is less than or equal the maximum number of messages to sync.  This should
results in message synchronization not taking an inordinate amount of time, especially for users
with a large number of messages in their mailboxes.